Use hard time limit on gradual memory reclaim
This is a purely theoretical implementation of the fix proposed in https://github.com/Artoria2e5, under the assumption that an incomplete reclaim is causing things to throttle and hang. It is completely untested nor do I have any intention of testing it. For CLA purposes, I disclaim any rights that I may or may not have on this code.